Trying to get CI working by making PowerArgs a submodule.
Replace PowerArgs references with submodule project refs

Updated project files to replace direct references to the PowerArgs library with project references to a submodule version located in the `external` directory. Modified `.gitmodules` to include the PowerArgs submodule and updated the GitHub Actions workflow in `dotnet.yml` to ensure submodules are checked out during the build process.